Fix missing DTO compile include and CorreoService source path
